LOAD_POWER not necessary an integer.
After recent update of xiaomi firmware, there's "0.00" shows in LOAD_POWER, which suggests that this field is not necessary an integer.

+++ components/switch/xiaomi.py
@@ -71,7 +71,8 @@ class XiaomiGenericSwitch(XiaomiDevice, SwitchDevice):
                  ATTR_LOAD_POWER: self._load_power,
                  ATTR_POWER_CONSUMED: self._power_consumed}
         attrs.update(super().device_state_attributes)
-        return attrs
+        return attrs
+
     def turn_on(self, **kwargs):
         """Turn the switch on."""
         self.xiaomi_hub.write_to_hub(self._sid, self._data_key, 'on')
@@ -81,7 +82,8 @@ class XiaomiGenericSwitch(XiaomiDevice, SwitchDevice):
         self.xiaomi_hub.write_to_hub(self._sid, self._data_key, 'off')
 
     def parse_data(self, data):
-        """Parse data sent by gateway"""
+        """Parse data sent by gateway"""
+
         if IN_USE in data:
             self._in_use = int(data[IN_USE])
             if not self._in_use:
@@ -91,7 +93,7 @@ class XiaomiGenericSwitch(XiaomiDevice, SwitchDevice):
             self._power_consumed = int(data[POWER_CONSUMED])
 
         if LOAD_POWER in data:
-            self._load_power = int(data[LOAD_POWER])
+            self._load_power = float(data[LOAD_POWER])
 
         value = data.get(self._data_key)
         if value is None:
